osx: fix new compiler errors after rebase
Must explicitly cast zfix to an int in subscr.cpp. I think this is
related to 64bit change. Anyhow, both sides of ternary operators
must be the same type and so best to be explicit with the conversion.

Trivial fix in parser header, clang doesn't like unnecessary
qualifications.

+++ src/subscr.cpp
@@ -4058,8 +4058,8 @@ void show_custom_subscreen(BITMAP *dest, miscQdata *misc, subscreen_group *css, 
 						dh = oldsel ? (tempsel->extend > 2 ? tempsel->txsz*16 : 16) : ((tmpitm.overrideFLAGS & itemdataOVERRIDE_HIT_HEIGHT) ? tmpitm.hysz : 16);
 					int32_t sxofs = oldsel ? 0 : (tempsel->extend > 2 ? tempsel->hxofs : 0),
 						syofs = oldsel ? 0 : (tempsel->extend > 2 ? tempsel->hyofs : 0),
-						dxofs = oldsel ? (tempsel->extend > 2 ? tempsel->xofs : 0) : ((tmpitm.overrideFLAGS & itemdataOVERRIDE_HIT_X_OFFSET) ? tmpitm.hxofs : 0) + (tempsel->extend > 2 ? tempsel->xofs : 0),
-						dyofs = oldsel ? (tempsel->extend > 2 ? tempsel->yofs : 0) : ((tmpitm.overrideFLAGS & itemdataOVERRIDE_HIT_Y_OFFSET) ? tmpitm.hyofs : 0) + (tempsel->extend > 2 ? tempsel->yofs : 0);
+						dxofs = oldsel ? (tempsel->extend > 2 ? (int)tempsel->xofs : 0) : ((tmpitm.overrideFLAGS & itemdataOVERRIDE_HIT_X_OFFSET) ? tmpitm.hxofs : 0) + (tempsel->extend > 2 ? (int)tempsel->xofs : 0),
+						dyofs = oldsel ? (tempsel->extend > 2 ? (int)tempsel->yofs : 0) : ((tmpitm.overrideFLAGS & itemdataOVERRIDE_HIT_Y_OFFSET) ? tmpitm.hyofs : 0) + (tempsel->extend > 2 ? (int)tempsel->yofs : 0);
 					BITMAP* tmpbmp = create_bitmap_ex(8,sw,sh);
 					for(int32_t j=0; j<4; ++j)
 					{
